UPDATE: delete persist-store and some fixing

+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Route, Routes, useLocation } from 'react-router-dom';
 import { lazy, Suspense, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { nanoid } from 'nanoid';
+import { useDispatch, useSelector } from 'react-redux';
+import { customAlphabet } from 'nanoid';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
@@ -15,15 +15,20 @@ const ChatPage = lazy(() => import('./pages/ChatPage'));
 const NotFoundPage = lazy(() => import('./pages/NotFoundPage'));
 
 import { authUser } from './redux/chat/operations';
+import { selectChat } from './redux/chat/selectors';
 
 export const App = () => {
   const location = useLocation();
   const dispatch = useDispatch();
+  const existingChat = useSelector(selectChat);
 
   useEffect(() => {
+    if (existingChat._id) return;
+
     const searchParams = new URLSearchParams(location.search);
     const userIdParam = searchParams.get('userId');
-    const storedUserId = localStorage.getItem('userId') || nanoid(24);
+    const storedUserId =
+      localStorage.getItem('userId') || customAlphabet('0123456789', 24)();
 
     if (userIdParam) {
       localStorage.setItem('userId', userIdParam);
@@ -32,7 +37,7 @@ export const App = () => {
       localStorage.setItem('userId', storedUserId);
       dispatch(authUser(storedUserId));
     }
-  }, [dispatch, location.search]);
+  }, [dispatch, existingChat._id, location.search]);
 
   return (
     <>
